// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "packages".
 */
export interface Package {
  id: number;
  destination?: string | null;
  title?: string | null;
  package?: string | null;
  cost?: number | null;
  image?: (number | null) | Media;
  is_popular?: boolean | null;
  popular_order?: number | null;
  order?: number | null;
  type?: ("domestic" | "international" | "cruise") | null;
  category?: (number | Category)[] | null;
  location?: (number | null) | Location;
  gallery?: (number | Media)[] | null;
  overview?: string | null;
  highlights?: {
    chips?:
      | {
          chip?: string | null;
          id?: string | null;
        }[]
      | null;
    pointers?:
      | {
          pointer?: string | null;
          id?: string | null;
        }[]
      | null;
  };
  day_breakup?: {
    days?:
      | {
          title?: string | null;
          break_up?:
            | {
                title?: string | null;
                subtitle?: string | null;
                id?: string | null;
              }[]
            | null;
          id?: string | null;
        }[]
      | null;
  };
  accomodations?: {
    locations?:
      | {
          image?: (number | null) | Media;
          name?: string | null;
          location?: string | null;
          type?: string | null;
          whats_included?: string | null;
          link?: string | null;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "locations".
 */
export interface Location {
  id: number;
  name: string;
  updatedAt: string;
  createdAt: string;
}